001_Chatbot/backend/requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
requests
langchain
langchain-core
langchain-community